type experiment
putting the learning rate, tau and the update cadence back to the values from udacity's project, leaving batch size large as the only change. This is because I noticed one other difference from the DQN code and mine and that is unity was not accepting my numpy int64 type actions, and so I had casst them to int32s before returning them from the step function, when it turns out I could have just done it right in time to send it to the unity enviornment and left them as numpy int64 values the rest of the time. this doesn't fix my issue but it made the scores higher on average I think once again, like it is finding a better local minima to get stuck in.  now maybe if I combine this with more experience memory or a different value for tau or some other tweak to the hyperparameters.

+++ reinforcement_learning/policy.py
@@ -22,11 +22,9 @@ def policy_function(epsilon, q_network, state, action_size):
     greater_than_epsilon = lambda epsilon: random.random() > epsilon
     # have to cast the argmax values as numpy int32, otherwise unity mlagents code breaks environment.py line 322 looking for 'keys'
 
-    max_action = lambda *args: np.argmax(action_values.cpu().data.numpy()).astype(
-        np.int32
-    )
+    max_action = lambda *args: np.argmax(action_values.cpu().data.numpy())
 
-    random_action = lambda *args: random.choice(np.arange(action_size)).astype(np.int32)
+    random_action = lambda *args: random.choice(np.arange(action_size))
 
     return (greater_than_epsilon(epsilon) and max_action()) or random_action()
